Split controllers into individual files
